Rename Nationality page to Citizenship
To reflect the new wording of the questions
